moving camera options to mvg
