+++ arrows/mvg/CMakeLists.txt
@@ -8,7 +8,7 @@ set( plugin_mvg_headers
   algo/initialize_cameras_landmarks.h
   algo/initialize_cameras_landmarks_basic.h
   algo/triangulate_landmarks.h
-
+  camera_options.h
   epipolar_geometry.h
   interpolate_camera.h
   necker_reverse.h
@@ -35,7 +35,7 @@ set( plugin_mvg_sources
   algo/initialize_cameras_landmarks.cxx
   algo/initialize_cameras_landmarks_basic.cxx
   algo/triangulate_landmarks.cxx
-
+  camera_options.cxx
   epipolar_geometry.cxx
   interpolate_camera.cxx
   necker_reverse.cxx
